Moved categories to activity level
